Remove EmbeddedAssemblyGenerator
Since GTFO uses Newtonsoft.Json, Unhollower creates an unhollowed
(but stripped) version of the library. The embedded assembly generator
would rely on the AppDomain's AssemblyResolve event to supply missing
libraries. This is however impossible to do with Newtonsoft.Json since
BepInEx loads all of the unhollowed assemblies in their mono AppDomain,
causing Newtonsoft.Json to resolve to the unhollowed assembly.

For now, ILRepack seems to be the best alternative. Unfortunately all of
the ILRepack MSBuild libraries seem to be broken; manual merging of the
assemblies is required.